#3 use generic func for parse json
