+++ HomeForSale/Pages/Homes/HomesInfoTableCoordinator.swift
@@ -15,7 +15,7 @@ final class HomesInfoTableCoordinator : Coordinator {
     }
     
     func start() {
-        let dataManager = NetworkDataManager(testMode: true)
+        let dataManager = DataManager(testMode: true)
         let viewModel = HomesViewModel(dataManager: dataManager)
         let viewController = HomesViewController()
         viewController.viewModel = viewModel
